requirements.txt
APScheduler==3.5.3
kafka-python==1.4.4